+++ src/services/CommitDiscussions.js
@@ -1,0 +1,9 @@
+import { ResourceDiscussions } from '../templates';
+
+class CommitDiscussions extends ResourceDiscussions {
+  constructor(options) {
+    super('projects', 'commits', options);
+  }
+}
+
+export default CommitDiscussions;